requirements.txt
streamlit
pandas
python-dotenv
requests
beautifulsoup4
numpy
plotly
PyYAML
pytest
xradar
fsspec
s3fs
cartopy
matplotlib
mlflow
scikit-learn
torch
transformers
accelerate